has errors
book.php, author.php errors

+++ app/Http/Controllers/AuthorController.php
@@ -15,6 +15,7 @@ class AuthorController extends Controller
     public function index()
     {
         $authors = Author::all();
+        //dd($authors);
        return view('author.index', ['authors' => $authors]);
     }
 
@@ -38,10 +39,10 @@ class AuthorController extends Controller
     public function store(Request $request)
     {
     $author = new Author;
-$author->name = $request->author_name;
-$author->surname = $request->author_surname;
-$author->save();
-return redirect()->route('author.index');
+    $author->name = $request->author_name;
+    $author->surname = $request->author_surname;
+    $author->save();
+    return redirect()->route('author.index');
 
     }
 
@@ -90,7 +91,11 @@ return redirect()->route('author.index');
      */
     public function destroy(Author $author)
     {
+        if($author->authorBooks->count()){
+            return 'Trinti negalima, nes turi knygų';
+        }
         $author->delete();
-       return redirect()->route('author.index');
+        return redirect()->route('author.index');
+ 
     }
 }
